requirements.txt
customtkinter
pymupdf
packaging
Pillow
